refactor: removal of repeated lines from DefaultFooter
